rafs: refine builder Bootstrap implementation
Refine builder Bootstrap implementation for maintenance.
