+++ rafs/src/builder/core/bootstrap.rs
@@ -1,41 +1,29 @@
 // Copyright 2020 Ant Group. All rights reserved.
+// Copyright (C) 2023 Alibaba Cloud. All rights reserved.
 //
 // SPDX-License-Identifier: Apache-2.0
 
-use std::collections::BTreeMap;
-use std::convert::TryFrom;
 use std::ffi::OsString;
-use std::io::SeekFrom;
-use std::mem::size_of;
 
-use anyhow::{bail, Context, Error, Result};
-use nydus_storage::device::BlobFeatures;
-use nydus_utils::digest::{self, DigestHasher, RafsDigest};
+use anyhow::{Context, Error, Result};
+use nydus_utils::digest::{self, RafsDigest};
 use nydus_utils::{root_tracer, timing_tracer};
 
 use super::node::Node;
 use super::overlay::{WhiteoutType, OVERLAYFS_WHITEOUT_OPAQUE};
 use crate::builder::{
-    ArtifactStorage, BlobManager, BootstrapContext, BootstrapManager, BuildContext, ConversionType,
-    Tree,
-};
-use crate::metadata::layout::v5::{
-    RafsV5BlobTable, RafsV5ChunkInfo, RafsV5InodeTable, RafsV5SuperBlock, RafsV5XAttrsTable,
-};
-use crate::metadata::layout::v6::{
-    align_offset, calculate_nid, RafsV6BlobTable, RafsV6Device, RafsV6SuperBlock,
-    RafsV6SuperBlockExt, EROFS_BLOCK_SIZE, EROFS_DEVTABLE_OFFSET, EROFS_INODE_SLOT_SIZE,
+    ArtifactStorage, BlobManager, BootstrapContext, BootstrapManager, BuildContext, Tree,
 };
 use crate::metadata::layout::{RafsBlobTable, RAFS_V5_ROOT_INODE};
-use crate::metadata::{RafsStore, RafsSuper, RafsSuperConfig};
+use crate::metadata::{RafsSuper, RafsSuperConfig};
 
 pub(crate) const STARGZ_DEFAULT_BLOCK_SIZE: u32 = 4 << 20;
-const WRITE_PADDING_DATA: [u8; 4096] = [0u8; 4096];
 
+/// RAFS bootstrap/meta blob builder.
 pub struct Bootstrap {}
 
 impl Bootstrap {
-    /// Create a new instance of `Bootstrap`.
+    /// Create a new instance of [Bootstrap].
     pub fn new() -> Result<Self> {
         Ok(Self {})
     }
@@ -51,7 +39,7 @@ impl Bootstrap {
     /// - files/directories to be removed from the lower layer, at the head of the array. The order
     ///   of removal operations are bottom-up, that means child files/directories is in front of its
     ///   parent.
-    /// - files/directories to added/modified in into the lower layer, at the tail of the array.
+    /// - files/directories to added/modified into the lower layer, at the tail of the array.
     ///   The order of addition/modification operations are top-down, that means directories is
     ///   ahead its children.
     ///
@@ -60,7 +48,7 @@ impl Bootstrap {
         &mut self,
         ctx: &mut BuildContext,
         bootstrap_ctx: &mut BootstrapContext,
-        tree: &mut Tree,
+        mut tree: Tree,
     ) -> Result<()> {
         // used to compute nid(ino) for v6
         let root_offset = bootstrap_ctx.offset;
@@ -77,10 +65,10 @@ impl Bootstrap {
         ctx.prefetch.insert_if_need(&tree.node);
         nodes.push(tree.node.clone());
 
-        self.build_rafs(ctx, bootstrap_ctx, tree, &mut nodes)?;
+        Self::build_rafs(ctx, bootstrap_ctx, &mut tree, &mut nodes)?;
         if ctx.fs_version.is_v6() && !bootstrap_ctx.layered {
             // generate on-disk metadata layout for v6
-            self.rafsv6_update_dirents(&mut nodes, tree, root_offset);
+            Self::v6_update_dirents(&mut nodes, &tree, root_offset);
         }
         bootstrap_ctx.nodes = nodes;
 
@@ -128,11 +116,34 @@ impl Bootstrap {
         Ok(tree)
     }
 
-    #[allow(clippy::only_used_in_recursion)]
+    /// Dump the RAFS filesystem meta information to meta blob.
+    pub fn dump(
+        &mut self,
+        ctx: &mut BuildContext,
+        bootstrap_storage: &mut Option<ArtifactStorage>,
+        bootstrap_ctx: &mut BootstrapContext,
+        blob_table: &RafsBlobTable,
+    ) -> Result<()> {
+        match blob_table {
+            RafsBlobTable::V5(table) => self.v5_dump(ctx, bootstrap_ctx, table)?,
+            RafsBlobTable::V6(table) => self.v6_dump(ctx, bootstrap_ctx, table)?,
+        }
+
+        if let Some(ArtifactStorage::FileDir(p)) = bootstrap_storage {
+            let bootstrap_data = bootstrap_ctx.writer.as_bytes()?;
+            let digest = RafsDigest::from_buf(&bootstrap_data, digest::Algorithm::Sha256);
+            let name = digest.to_string();
+            bootstrap_ctx.writer.finalize(Some(name.clone()))?;
+            *bootstrap_storage = Some(ArtifactStorage::SingleFile(p.join(name)));
+            Ok(())
+        } else {
+            bootstrap_ctx.writer.finalize(Some(String::default()))
+        }
+    }
+
     /// Traverse node tree, set inode index, ino, child_index and child_count etc according to the
     /// RAFS metadata format, then store to nodes collection.
     fn build_rafs(
-        &mut self,
         ctx: &mut BuildContext,
         bootstrap_ctx: &mut BootstrapContext,
         tree: &mut Tree,
@@ -140,6 +151,7 @@ impl Bootstrap {
     ) -> Result<()> {
         let index = nodes.len() as u32 + 1;
         let parent = &mut nodes[tree.node.index as usize - 1];
+        let parent_ino = parent.inode.ino();
 
         // Maybe the parent is not a directory in multi-layers build scenario, so we check here.
         if parent.is_dir() {
@@ -156,22 +168,19 @@ impl Bootstrap {
 
         if ctx.fs_version.is_v6() {
             tree.node.v6_offset = parent.v6_offset;
-            // alignment for inode, which is 32 bytes;
-            bootstrap_ctx.align_offset(EROFS_INODE_SLOT_SIZE as u64);
         }
 
         // Cache dir tree for BFS walk
         let mut dirs: Vec<&mut Tree> = Vec::new();
-        let parent_ino = parent.inode.ino();
-
         for child in tree.children.iter_mut() {
             let index = nodes.len() as u64 + 1;
             child.node.index = index;
             child.node.inode.set_parent(parent_ino);
 
-            // Hardlink handle, all hardlink nodes' ino, nlink should be the same,
-            // because the real_ino may be conflicted between different layers,
-            // so we need to find hardlink node index list in the layer where the node is located.
+            // Handle hardlink.
+            // All hardlink nodes' ino and nlink should be the same.
+            // We need to find hardlink node index list in the layer where the node is located
+            // because the real_ino may be different among different layers,
             let mut v6_hardlink_offset: Option<u64> = None;
             if let Some(indexes) = bootstrap_ctx.inode_map.get_mut(&(
                 child.node.layer_idx,
@@ -208,7 +217,6 @@ impl Bootstrap {
                 child
                     .node
                     .v6_set_offset(bootstrap_ctx, v6_hardlink_offset)?;
-                bootstrap_ctx.align_offset(EROFS_INODE_SLOT_SIZE as u64);
             }
 
             // Store node for bootstrap & blob dump.
@@ -240,9 +248,7 @@ impl Bootstrap {
                     }
                     nodes.push(child.node.clone());
                 }
-                _ => {
-                    nodes.push(child.node.clone());
-                }
+                _ => nodes.push(child.node.clone()),
             }
 
             ctx.prefetch.insert_if_need(&child.node);
@@ -257,61 +263,15 @@ impl Bootstrap {
         // updating parent directory's nlink here is reliable since builder re-constructs
         // the entire tree and intends to layout all inodes into a plain array fetching
         // from the previously applied tree.
-        let parent_dir = &mut nodes[tree.node.index as usize - 1];
-        parent_dir.inode.set_nlink((2 + dirs.len()) as u32);
-
+        let parent = &mut nodes[tree.node.index as usize - 1];
+        if parent.is_dir() {
+            parent.inode.set_nlink((2 + dirs.len()) as u32);
+        }
         for dir in dirs {
-            self.build_rafs(ctx, bootstrap_ctx, dir, nodes)?;
+            Self::build_rafs(ctx, bootstrap_ctx, dir, nodes)?;
         }
 
         Ok(())
-    }
-
-    #[allow(clippy::only_used_in_recursion)]
-    /// Rafsv6 update offset
-    fn rafsv6_update_dirents(&self, nodes: &mut Vec<Node>, tree: &mut Tree, parent_offset: u64) {
-        let node = &mut nodes[tree.node.index as usize - 1];
-        let node_offset = node.v6_offset;
-        if !node.is_dir() {
-            return;
-        }
-
-        // dot & dotdot
-        // Type of libc::S_IFDIR is u16 on macos, so it need a conversion
-        // but compiler will report useless conversion on linux platform,
-        // so we add an allow annotation here.
-        #[allow(clippy::useless_conversion)]
-        {
-            node.v6_dirents
-                .push((node_offset, OsString::from("."), libc::S_IFDIR.into()));
-            node.v6_dirents
-                .push((parent_offset, OsString::from(".."), libc::S_IFDIR.into()));
-        }
-
-        let mut dirs: Vec<&mut Tree> = Vec::new();
-        for child in tree.children.iter_mut() {
-            trace!(
-                "{:?} child {:?} offset {}, mode {}",
-                node.name(),
-                child.node.name(),
-                child.node.v6_offset,
-                child.node.inode.mode()
-            );
-            node.v6_dirents.push((
-                child.node.v6_offset,
-                child.node.name().to_os_string(),
-                child.node.inode.mode(),
-            ));
-            if child.node.is_dir() {
-                dirs.push(child);
-            }
-        }
-        node.v6_dirents
-            .sort_unstable_by(|a, b| a.1.as_os_str().cmp(b.1.as_os_str()) as std::cmp::Ordering);
-
-        for dir in dirs {
-            self.rafsv6_update_dirents(nodes, dir, node_offset);
-        }
     }
 
     fn load_parent_bootstrap(
@@ -346,449 +306,5 @@ impl Bootstrap {
             .context("failed to build tree from bootstrap")?;
 
         Ok(tree)
-    }
-
-    /// Calculate inode digest for directory.
-    fn rafsv5_digest_node(
-        &self,
-        ctx: &mut BuildContext,
-        bootstrap_ctx: &mut BootstrapContext,
-        index: usize,
-    ) {
-        let node = &bootstrap_ctx.nodes[index];
-
-        // We have set digest for non-directory inode in the previous dump_blob workflow.
-        if node.is_dir() {
-            let child_index = node.inode.child_index();
-            let child_count = node.inode.child_count();
-            let mut inode_hasher = RafsDigest::hasher(ctx.digester);
-
-            for idx in child_index..child_index + child_count {
-                let child = &bootstrap_ctx.nodes[(idx - 1) as usize];
-                inode_hasher.digest_update(child.inode.digest().as_ref());
-            }
-
-            bootstrap_ctx.nodes[index]
-                .inode
-                .set_digest(inode_hasher.digest_finalize());
-        }
-    }
-
-    pub fn dump(
-        &mut self,
-        ctx: &mut BuildContext,
-        bootstrap_storage: &mut Option<ArtifactStorage>,
-        bootstrap_ctx: &mut BootstrapContext,
-        blob_table: &RafsBlobTable,
-    ) -> Result<()> {
-        match blob_table {
-            RafsBlobTable::V5(table) => self.rafsv5_dump(ctx, bootstrap_ctx, table)?,
-            RafsBlobTable::V6(table) => self.rafsv6_dump(ctx, bootstrap_ctx, table)?,
-        }
-
-        if let Some(ArtifactStorage::FileDir(p)) = bootstrap_storage {
-            let bootstrap_data = bootstrap_ctx.writer.as_bytes()?;
-            let digest = RafsDigest::from_buf(&bootstrap_data, digest::Algorithm::Sha256);
-            let name = digest.to_string();
-            bootstrap_ctx.writer.finalize(Some(name.clone()))?;
-            *bootstrap_storage = Some(ArtifactStorage::SingleFile(p.join(name)));
-            Ok(())
-        } else {
-            bootstrap_ctx.writer.finalize(Some(String::default()))
-        }
-    }
-
-    /// Dump bootstrap and blob file, return (Vec<blob_id>, blob_size)
-    fn rafsv5_dump(
-        &mut self,
-        ctx: &mut BuildContext,
-        bootstrap_ctx: &mut BootstrapContext,
-        blob_table: &RafsV5BlobTable,
-    ) -> Result<()> {
-        // Set inode digest, use reverse iteration order to reduce repeated digest calculations.
-        for idx in (0..bootstrap_ctx.nodes.len()).rev() {
-            self.rafsv5_digest_node(ctx, bootstrap_ctx, idx);
-        }
-
-        // Set inode table
-        let super_block_size = size_of::<RafsV5SuperBlock>();
-        let inode_table_entries = bootstrap_ctx.nodes.len() as u32;
-        let mut inode_table = RafsV5InodeTable::new(inode_table_entries as usize);
-        let inode_table_size = inode_table.size();
-
-        // Set prefetch table
-        let (prefetch_table_size, prefetch_table_entries) = if let Some(prefetch_table) =
-            ctx.prefetch.get_rafsv5_prefetch_table(&bootstrap_ctx.nodes)
-        {
-            (prefetch_table.size(), prefetch_table.len() as u32)
-        } else {
-            (0, 0u32)
-        };
-
-        // Set blob table, use sha256 string (length 64) as blob id if not specified
-        let prefetch_table_offset = super_block_size + inode_table_size;
-        let blob_table_offset = prefetch_table_offset + prefetch_table_size;
-        let blob_table_size = blob_table.size();
-        let extended_blob_table_offset = blob_table_offset + blob_table_size;
-        let extended_blob_table_size = blob_table.extended.size();
-        let extended_blob_table_entries = blob_table.extended.entries();
-
-        // Set super block
-        let mut super_block = RafsV5SuperBlock::new();
-        let inodes_count = bootstrap_ctx.inode_map.len() as u64;
-        super_block.set_inodes_count(inodes_count);
-        super_block.set_inode_table_offset(super_block_size as u64);
-        super_block.set_inode_table_entries(inode_table_entries);
-        super_block.set_blob_table_offset(blob_table_offset as u64);
-        super_block.set_blob_table_size(blob_table_size as u32);
-        super_block.set_extended_blob_table_offset(extended_blob_table_offset as u64);
-        super_block.set_extended_blob_table_entries(u32::try_from(extended_blob_table_entries)?);
-        super_block.set_prefetch_table_offset(prefetch_table_offset as u64);
-        super_block.set_prefetch_table_entries(prefetch_table_entries);
-        super_block.set_compressor(ctx.compressor);
-        super_block.set_digester(ctx.digester);
-        super_block.set_chunk_size(ctx.chunk_size);
-        if ctx.explicit_uidgid {
-            super_block.set_explicit_uidgid();
-        }
-        if ctx.conversion_type == ConversionType::EStargzIndexToRef {
-            super_block.set_block_size(STARGZ_DEFAULT_BLOCK_SIZE);
-        }
-
-        // Set inodes and chunks
-        let mut inode_offset = (super_block_size
-            + inode_table_size
-            + prefetch_table_size
-            + blob_table_size
-            + extended_blob_table_size) as u32;
-
-        let mut has_xattr = false;
-        for node in &mut bootstrap_ctx.nodes {
-            inode_table.set(node.index, inode_offset)?;
-            // Add inode size
-            inode_offset += node.inode.inode_size() as u32;
-            if node.inode.has_xattr() {
-                has_xattr = true;
-                if !node.info.xattrs.is_empty() {
-                    inode_offset += (size_of::<RafsV5XAttrsTable>()
-                        + node.info.xattrs.aligned_size_v5())
-                        as u32;
-                }
-            }
-            // Add chunks size
-            if node.is_reg() {
-                inode_offset += node.inode.child_count() * size_of::<RafsV5ChunkInfo>() as u32;
-            }
-        }
-        if has_xattr {
-            super_block.set_has_xattr();
-        }
-
-        // Dump super block
-        super_block
-            .store(bootstrap_ctx.writer.as_mut())
-            .context("failed to store superblock")?;
-
-        // Dump inode table
-        inode_table
-            .store(bootstrap_ctx.writer.as_mut())
-            .context("failed to store inode table")?;
-
-        // Dump prefetch table
-        if let Some(mut prefetch_table) =
-            ctx.prefetch.get_rafsv5_prefetch_table(&bootstrap_ctx.nodes)
-        {
-            prefetch_table
-                .store(bootstrap_ctx.writer.as_mut())
-                .context("failed to store prefetch table")?;
-        }
-
-        // Dump blob table
-        blob_table
-            .store(bootstrap_ctx.writer.as_mut())
-            .context("failed to store blob table")?;
-
-        // Dump extended blob table
-        blob_table
-            .store_extended(bootstrap_ctx.writer.as_mut())
-            .context("failed to store extended blob table")?;
-
-        // Dump inodes and chunks
-        timing_tracer!(
-            {
-                for node in &bootstrap_ctx.nodes {
-                    node.dump_bootstrap_v5(ctx, bootstrap_ctx.writer.as_mut())
-                        .context("failed to dump bootstrap")?;
-                }
-
-                Ok(())
-            },
-            "dump_bootstrap",
-            Result<()>
-        )?;
-
-        Ok(())
-    }
-
-    /// Dump bootstrap and blob file, return (Vec<blob_id>, blob_size)
-    fn rafsv6_dump(
-        &mut self,
-        ctx: &mut BuildContext,
-        bootstrap_ctx: &mut BootstrapContext,
-        blob_table: &RafsV6BlobTable,
-    ) -> Result<()> {
-        // Rafs v6 disk layout
-        //
-        //  EROFS_SUPER_OFFSET
-        //     |
-        // +---+---------+------------+-------------+----------------------------------------------+
-        // |   |         |            |             |                 |         |                  |
-        // |1k |super    |extended    | blob table  |  prefetch table | inodes  | chunk info table |
-        // |   |block    |superblock+ |             |                 |         |                  |
-        // |   |         |devslot     |             |                 |         |                  |
-        // +---+---------+------------+-------------+----------------------------------------------+
-
-        let blobs = blob_table.get_all();
-        let devtable_len = blobs.len() * size_of::<RafsV6Device>();
-        let blob_table_size = blob_table.size() as u64;
-        let blob_table_offset = align_offset(
-            (EROFS_DEVTABLE_OFFSET as u64) + devtable_len as u64,
-            EROFS_BLOCK_SIZE as u64,
-        );
-        let blob_table_entries = blobs.len();
-        assert!(blob_table_entries < u8::MAX as usize);
-        trace!(
-            "devtable len {} blob table offset {} blob table size {}",
-            devtable_len,
-            blob_table_offset,
-            blob_table_size
-        );
-
-        let (prefetch_table_offset, prefetch_table_size) =
-            // If blob_table_size equal to 0, there is no prefetch.
-            if ctx.prefetch.fs_prefetch_rule_count() > 0 && blob_table_size > 0 {
-                // Prefetch table is very close to blob devices table
-                let offset = blob_table_offset + blob_table_size;
-                // Each prefetched file has is nid of `u32` filled into prefetch table.
-                let size = ctx.prefetch.fs_prefetch_rule_count() * size_of::<u32>() as u32;
-                trace!("prefetch table locates at offset {} size {}", offset, size);
-                (offset, size)
-            } else {
-                (0, 0)
-            };
-
-        // Make the superblock's meta_blkaddr one block ahead of the inode table,
-        // to avoid using 0 as root nid.
-        // inode offset = meta_blkaddr * block_size + 32 * nid
-        // When using nid 0 as root nid,
-        // the root directory will not be shown by glibc's getdents/readdir.
-        // Because in some OS, ino == 0 represents corresponding file is deleted.
-        let orig_meta_addr = bootstrap_ctx.nodes[0].v6_offset - EROFS_BLOCK_SIZE;
-        let meta_addr = if blob_table_size > 0 {
-            align_offset(
-                blob_table_offset + blob_table_size + prefetch_table_size as u64,
-                EROFS_BLOCK_SIZE as u64,
-            )
-        } else {
-            orig_meta_addr
-        };
-
-        // get devt_slotoff
-        let root_nid = calculate_nid(
-            bootstrap_ctx.nodes[0].v6_offset + (meta_addr - orig_meta_addr),
-            meta_addr,
-        );
-
-        // Prepare extended super block
-        let ext_sb_offset = bootstrap_ctx.writer.seek_current(0)?;
-        let mut ext_sb = RafsV6SuperBlockExt::new();
-        ext_sb.set_compressor(ctx.compressor);
-        ext_sb.set_digester(ctx.digester);
-        ext_sb.set_chunk_size(ctx.chunk_size);
-        ext_sb.set_blob_table_offset(blob_table_offset);
-        ext_sb.set_blob_table_size(blob_table_size as u32);
-
-        // collect all chunks in this bootstrap.
-        // HashChunkDict cannot be used here, because there will be duplicate chunks between layers,
-        // but there is no deduplication during the actual construction.
-        // Each layer uses the corresponding chunk in the blob of its own layer.
-        // If HashChunkDict is used here, it will cause duplication. The chunks are removed,
-        // resulting in incomplete chunk info.
-        let mut chunk_cache = BTreeMap::new();
-
-        // Dump bootstrap
-        timing_tracer!(
-            {
-                for node in &mut bootstrap_ctx.nodes {
-                    node.dump_bootstrap_v6(
-                        ctx,
-                        bootstrap_ctx.writer.as_mut(),
-                        orig_meta_addr,
-                        meta_addr,
-                        &mut chunk_cache,
-                    )
-                    .context("failed to dump bootstrap")?;
-                }
-
-                Ok(())
-            },
-            "dump_bootstrap",
-            Result<()>
-        )?;
-        Self::rafsv6_align_to_block(bootstrap_ctx)?;
-
-        // `Node` offset might be updated during above inodes dumping. So `get_prefetch_table` after it.
-        let prefetch_table = ctx
-            .prefetch
-            .get_rafsv6_prefetch_table(&bootstrap_ctx.nodes, meta_addr);
-        if let Some(mut pt) = prefetch_table {
-            assert!(pt.len() * size_of::<u32>() <= prefetch_table_size as usize);
-            // Device slots are very close to extended super block.
-            ext_sb.set_prefetch_table_offset(prefetch_table_offset);
-            ext_sb.set_prefetch_table_size(prefetch_table_size);
-            bootstrap_ctx
-                .writer
-                .seek_offset(prefetch_table_offset as u64)
-                .context("failed seek prefetch table offset")?;
-            pt.store(bootstrap_ctx.writer.as_mut()).unwrap();
-        }
-
-        // TODO: get rid of the chunk info array.
-        // Dump chunk info array.
-        let chunk_table_offset = bootstrap_ctx
-            .writer
-            .seek_to_end()
-            .context("failed to seek to bootstrap's end for chunk table")?;
-        let mut chunk_table_size: u64 = 0;
-        for (_, chunk) in chunk_cache.iter() {
-            let chunk_size = chunk
-                .store(bootstrap_ctx.writer.as_mut())
-                .context("failed to dump chunk table")?;
-            chunk_table_size += chunk_size as u64;
-        }
-        ext_sb.set_chunk_table(chunk_table_offset, chunk_table_size);
-        debug!(
-            "chunk_table offset {} size {}",
-            chunk_table_offset, chunk_table_size
-        );
-        Self::rafsv6_align_to_block(bootstrap_ctx)?;
-
-        // Prepare device slots.
-        let pos = bootstrap_ctx
-            .writer
-            .seek_to_end()
-            .context("failed to seek to bootstrap's end for chunk table")?;
-        assert_eq!(pos % EROFS_BLOCK_SIZE, 0);
-        let mut mapped_blkaddr = Self::align_mapped_blkaddr((pos / EROFS_BLOCK_SIZE) as u32);
-        let mut devtable: Vec<RafsV6Device> = Vec::new();
-        let mut block_count = 0u32;
-        let mut inlined_chunk_digest = true;
-        for entry in blobs.iter() {
-            let mut devslot = RafsV6Device::new();
-            // blob id is String, which is processed by sha256.finalize().
-            if entry.blob_id().is_empty() {
-                bail!(" blob id is empty");
-            } else if entry.blob_id().len() > 64 {
-                bail!(format!(
-                    "blob id length is bigger than 64 bytes, blob id {:?}",
-                    entry.blob_id()
-                ));
-            } else if entry.uncompressed_size() / EROFS_BLOCK_SIZE > u32::MAX as u64 {
-                bail!(format!(
-                    "uncompressed blob size (0x:{:x}) is too big",
-                    entry.uncompressed_size()
-                ));
-            }
-            if !entry.has_feature(BlobFeatures::INLINED_CHUNK_DIGEST) {
-                inlined_chunk_digest = false;
-            }
-            let cnt = (entry.uncompressed_size() / EROFS_BLOCK_SIZE) as u32;
-            assert!(block_count.checked_add(cnt).is_some());
-            block_count += cnt;
-            let id = entry.blob_id();
-            let id = id.as_bytes();
-            let mut blob_id = [0u8; 64];
-            blob_id[..id.len()].copy_from_slice(id);
-            devslot.set_blob_id(&blob_id);
-            devslot.set_blocks(cnt);
-            devslot.set_mapped_blkaddr(mapped_blkaddr);
-            devtable.push(devslot);
-
-            mapped_blkaddr = Self::align_mapped_blkaddr(mapped_blkaddr + cnt);
-        }
-
-        // Dump super block
-        let mut sb = RafsV6SuperBlock::new();
-        sb.set_inos(bootstrap_ctx.nodes.len() as u64);
-        sb.set_blocks(block_count);
-        sb.set_root_nid(root_nid as u16);
-        sb.set_meta_addr(meta_addr);
-        sb.set_extra_devices(blob_table_entries as u16);
-        bootstrap_ctx.writer.seek(SeekFrom::Start(0))?;
-        sb.store(bootstrap_ctx.writer.as_mut())
-            .context("failed to store SB")?;
-
-        // Dump extended super block.
-        if ctx.explicit_uidgid {
-            ext_sb.set_explicit_uidgid();
-        }
-        if ctx.has_xattr {
-            ext_sb.set_has_xattr();
-        }
-        if inlined_chunk_digest {
-            ext_sb.set_inlined_chunk_digest();
-        }
-        bootstrap_ctx.writer.seek(SeekFrom::Start(ext_sb_offset))?;
-        ext_sb
-            .store(bootstrap_ctx.writer.as_mut())
-            .context("failed to store extended super block")?;
-
-        // Dump device slots.
-        bootstrap_ctx
-            .writer
-            .seek_offset(EROFS_DEVTABLE_OFFSET as u64)
-            .context("failed to seek devtslot")?;
-        for slot in devtable.iter() {
-            slot.store(bootstrap_ctx.writer.as_mut())
-                .context("failed to store device slot")?;
-        }
-
-        // Dump blob table
-        bootstrap_ctx
-            .writer
-            .seek_offset(blob_table_offset as u64)
-            .context("failed seek for extended blob table offset")?;
-        blob_table
-            .store(bootstrap_ctx.writer.as_mut())
-            .context("failed to store extended blob table")?;
-
-        Ok(())
-    }
-
-    fn rafsv6_align_to_block(bootstrap_ctx: &mut BootstrapContext) -> Result<()> {
-        bootstrap_ctx
-            .writer
-            .flush()
-            .context("failed to flush bootstrap")?;
-        let pos = bootstrap_ctx
-            .writer
-            .seek_to_end()
-            .context("failed to seek to bootstrap's end for chunk table")?;
-        let padding = align_offset(pos, EROFS_BLOCK_SIZE as u64) - pos;
-        bootstrap_ctx
-            .writer
-            .write_all(&WRITE_PADDING_DATA[0..padding as usize])
-            .context("failed to write 0 to padding of bootstrap's end for chunk table")?;
-        bootstrap_ctx
-            .writer
-            .flush()
-            .context("failed to flush bootstrap")?;
-        Ok(())
-    }
-
-    fn align_mapped_blkaddr(addr: u32) -> u32 {
-        // TODO: define a const in nydus-service for 0x20_0000
-        let blocks = (0x20_0000u64 / EROFS_BLOCK_SIZE) as u32;
-        (addr + blocks - 1) / blocks * blocks
     }
 }
